--- .env
@@ -1,1 +1,0 @@
-NEXTAUTH_URL=http://localhost:3000
