+++ src/app/signIn/page.jsx
@@ -1,21 +1,23 @@
 "use client";
 import { useState } from "react";
-import { useCreateUserWithEmailAndPassword } from "react-firebase-hooks/auth";
-import { auth } from "@/app/firebase/config";
+import { useSignInWithEmailAndPassword } from "react-firebase-hooks/auth";
+import { auth } from "@/app/firebase";
+import { useRouter } from "next/navigation";
 
-const SignUp = () => {
+const SignIn = () => {
   const [email, setEmail] = useState("");
   const [password, setPassword] = useState("");
-  const [createUserWithEmailAndPassword] =
-    useCreateUserWithEmailAndPassword(auth);
+  const [signInWithEmailAndPassword] = useSignInWithEmailAndPassword(auth);
+  const router = useRouter();
 
-  const handleSignUp = async () => {
+  const handleSignIn = async () => {
     try {
-      const res = await createUserWithEmailAndPassword(email, password);
+      const res = await signInWithEmailAndPassword(email, password);
       console.log({ res });
       sessionStorage.setItem("user", true);
       setEmail("");
       setPassword("");
+      router.push("/");
     } catch (e) {
       console.error(e);
     }
@@ -24,7 +26,7 @@ const SignUp = () => {
   return (
     <div className="min-h-screen flex items-center justify-center bg-gray-900">
       <div className="bg-gray-800 p-10 rounded-lg shadow-xl w-96">
-        <h1 className="text-white text-2xl mb-5">Sign Up</h1>
+        <h1 className="text-white text-2xl mb-5">Sign In</h1>
         <input
           type="email"
           placeholder="Email"
@@ -40,13 +42,13 @@ const SignUp = () => {
           className="w-full p-3 mb-4 bg-gray-700 rounded outline-none text-white placeholder-gray-500"
         />
         <button
-          onClick={handleSignUp}
+          onClick={handleSignIn}
           className="w-full p-3 bg-indigo-600 rounded text-white hover:bg-indigo-500">
-          Sign Up
+          Sign In
         </button>
       </div>
     </div>
   );
 };
 
-export default SignUp;
+export default SignIn;
