+++ src/app/signup/page.jsx
@@ -1,23 +1,21 @@
 "use client";
 import { useState } from "react";
-import { useSignInWithEmailAndPassword } from "react-firebase-hooks/auth";
-import { auth } from "@/app/firebase/config";
-import { useRouter } from "next/navigation";
+import { useCreateUserWithEmailAndPassword } from "react-firebase-hooks/auth";
+import { auth } from "@/app/firebase";
 
-const SignIn = () => {
+const SignUp = () => {
   const [email, setEmail] = useState("");
   const [password, setPassword] = useState("");
-  const [signInWithEmailAndPassword] = useSignInWithEmailAndPassword(auth);
-  const router = useRouter();
+  const [createUserWithEmailAndPassword] =
+    useCreateUserWithEmailAndPassword(auth);
 
-  const handleSignIn = async () => {
+  const handleSignUp = async () => {
     try {
-      const res = await signInWithEmailAndPassword(email, password);
+      const res = await createUserWithEmailAndPassword(email, password);
       console.log({ res });
       sessionStorage.setItem("user", true);
       setEmail("");
       setPassword("");
-      router.push("/");
     } catch (e) {
       console.error(e);
     }
@@ -26,7 +24,7 @@ const SignIn = () => {
   return (
     <div className="min-h-screen flex items-center justify-center bg-gray-900">
       <div className="bg-gray-800 p-10 rounded-lg shadow-xl w-96">
-        <h1 className="text-white text-2xl mb-5">Sign In</h1>
+        <h1 className="text-white text-2xl mb-5">Sign Up</h1>
         <input
           type="email"
           placeholder="Email"
@@ -42,13 +40,13 @@ const SignIn = () => {
           className="w-full p-3 mb-4 bg-gray-700 rounded outline-none text-white placeholder-gray-500"
         />
         <button
-          onClick={handleSignIn}
+          onClick={handleSignUp}
           className="w-full p-3 bg-indigo-600 rounded text-white hover:bg-indigo-500">
-          Sign In
+          Sign Up
         </button>
       </div>
     </div>
   );
 };
 
-export default SignIn;
+export default SignUp;
